remove unnecessary files, change help output
