--- test/integration/secret-agent-conf.yaml
@@ -1,11 +1,0 @@
-service:
-  tcp:
-    endpoint: 0.0.0.0:3005
-
-secret-manager:
-  file:
-    resources:
-      "r1": "./secret-agent.json"
-
-log:
-  level: info
