--- test/integration/secret-agent.json
@@ -1,10 +1,0 @@
-{
-    "string_val": "c3Ry",
-    "int_val": "MTIzNA==",
-    "bool_val": "dHJ1ZQ==",
-    "compress_val": "enN0ZA==",
-    "encryption_val": "YWVzMjU2",
-    "s3_log_level_val": "ZGVidWc=",
-    "parallel_val": "MTY=",
-    "modified_by_val": "MjAxNi0wNS0xMl8wODoxMDozMA=="
-}
